[#671] refactor: 동아리 신청 배너 컴포넌트 통합
마이페이지, FAQ, 동아리 목록에서 각기 다른 스타일로 존재하던 배너를 shared/ui/ClubApplicationBanner 하나로 통합하고 디자인 스펙에 맞춤. 화살표 아이콘을 정사각형 lucide 아이콘 대신 8x12 전용 벡터로 교체해 왜곡 문제 수정.
