+++ src/widgets/club/ui/club-item-client-list.tsx
@@ -5,7 +5,7 @@ import useUniversityCode from '@/shared/hooks/useUniversityCode';
 import Link from 'next/link';
 
 import ClubItem from '@/entities/club/ui/club-item';
-import ClubApplicationBanner from './club-application-banner';
+import ClubApplicationBanner from '@/shared/ui/ClubApplicationBanner';
 import { Club } from '../model/type';
 
 const BANNER_INTERVAL = 4;
